+++ abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/infrastructure/configuration/PlatformModuleConfiguration.java
@@ -29,10 +29,10 @@ import static com.abixen.platform.service.businessintelligence.infrastructure.co
         com.abixen.platform.service.businessintelligence.multivisualisation.domain.repository.PackageMarker.class,
         com.abixen.platform.service.businessintelligence.multivisualisation.domain.service.PackageMarker.class,
         com.abixen.platform.service.businessintelligence.multivisualisation.application.service.PackageMarker.class,
-        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.impl.parser.PackageMarker.class,
-        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.impl.parser.preparer.PackageMarker.class,
-        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.impl.parser.reader.PackageMarker.class,
-        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.impl.parser.validator.PackageMarker.class,
+        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.parser.PackageMarker.class,
+        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.parser.impl.preparer.PackageMarker.class,
+        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.parser.impl.reader.PackageMarker.class,
+        com.abixen.platform.service.businessintelligence.multivisualisation.application.service.parser.impl.validator.PackageMarker.class,
         com.abixen.platform.service.businessintelligence.multivisualisation.interfaces.client.controller.PackageMarker.class,
         com.abixen.platform.service.businessintelligence.multivisualisation.interfaces.web.controller.PackageMarker.class,
         com.abixen.platform.service.businessintelligence.multivisualisation.application.converter.PackageMarker.class
