--- abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/multivisualisation/application/service/impl/FileParserFactory.java
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) 2010-present Abixen Systems. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-
-package com.abixen.platform.service.businessintelligence.multivisualisation.application.service.impl;
-
-import com.abixen.platform.service.businessintelligence.multivisualisation.application.service.FileDataParserService;
-
-public interface FileParserFactory {
-    FileDataParserService getParse(String extension);
-}
